budb: Remove unused globals
Remove the following unused struct fields and global variables:

- struct memoryDB.readTime
- src/budb/procs.c:lastOperation
- src/budb/procs.c:lastTrans

These have all be unused since OpenAFS 1.0; remove them. The global
'lastTrans', and field 'readTime' in struct memoryDB, are assigned a
value, but are otherwise never used.

Note that struct memoryDB is an in-memory representation of the buserver
database, not an on-disk format. (The field 'h', of type struct
dbHeader, is an on-disk struct, but we are not changing that.)

'lastOperation' and 'lastTrans' have the same names as global variables
in the kaserver, and the comments make kaserver references, so they may
have been copied from kaserver code long ago.

Change-Id: I7b76d43a19a9fdaef7d4b34991bcda83bb4a530d
Reviewed-on: https://gerrit.openafs.org/16460
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>