log: Remove kseal
kseal is a low-level test program that generates a token, given a
username and the pre-salted password for the AFS service key.  After
commit a6fe83d9a6 no-kseal-20070819, the 'kseal' test program was no
longer installed or built by default due to its limited usefulness.

Since that time it has bitrotted and thus no longer builds.
Remove it from the tree.

Change-Id: Ia281a36a15dedf3107caf4b88d6c9b0145ef3912
Reviewed-on: https://gerrit.openafs.org/16461
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>