bozo: Replace oldTime with oldExists in SaveOldFiles()
In the SaveOldFiles() function, the mtimes of the .OLD and .BAK files
are copied to a set of 32 bit integers. However, only the .BAK file
mtime is used in a time calculation.

- Add oldExists and bakExists flags to indicate when the files exist
  based on a successful stat() calls.

- Remove the oldTime variable which is no longer needed.

- Replace the file level BOZO_OLDTIME define with the local scope
  const SECONDS_PER_WEEK to clean up the code.

This commit removes a 32-bit integer used to represent time in a case we
did not actually use the time value.

Change-Id: Ib4c1b184f4af88eb5d96ad7d3edbe1d7b649fe36
Reviewed-on: https://gerrit.openafs.org/16306
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>