roken: Serialize librokenafs.la and librokenafs.a
Commit 48d181ca1f (libtool: Serialize building libfoo.la and libfoo.a)
serialized the steps for building several shared-only and static-only
libs, but missed the librokenafs library (probably because it uses the
$(LT_LDLIB_shlib_missing) makefile rule, instead of
$(LT_LDLIB_shlib_only)).

Serialize the steps for building the rokenafs libs like the others, to
make sure the same issue doesn't happen for librokenafs.

Thanks to kaduk@mit.edu for pointing out the missing case.

Change-Id: I57389e442e22dee7950c9cda5cc903a37f3b7126
Reviewed-on: https://gerrit.openafs.org/16303
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>