bozo: Remove BOZO_SAVE_CORES
Commit 7dd378687b9f (bozo-save-old-cores-by-date-option-20010212) added
some code to append a timestamp to bozo-processed core files, hidden
under the #ifdef BOZO_SAVE_CORES. Nothing ever sets BOZO_SAVE_CORES, so
using this functionality requires setting CFLAGS or changing
bozo/Makefile.

Commit 309122e86415
(configure-and-configure-libafs-use-acinclude-as-common-ground-20011008)
added a "todo" comment in acincludes.m4, likely as a reminder to make
this a configure time option.

Remove this, since currently it is almost never built and so will bit
rot. The functionality is a bit questionable, since it can easily fill
up the local disk if bosserver tries to restart a core-generating
process, since we never delete old core files. Some platforms also
provide their own core-managing functionality (e.g., abrtd,
systemd-coredump), and so this feature may be considered redundant.

If actually needed, it can be reintroduced as a bosserver run time
option in the future, ideally with some kind of rotation or limit to
avoid filling up the disk.

Change-Id: Ib417277276b28cf916b165782dd95b6f75ff14cb
Reviewed-on: https://gerrit.openafs.org/16305
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>