util: Remove redundant src/util/test/test_ktime.c test program
Commit 2e2494d6dbc3b (Add a sample test program to the new test suite)
converted the tests in this old program to C TAP tests, but left this
original program which was a test of the ktime_DateToLong() (a.k.a.
ktime_DateToInt32) function.

The old program is not built by default and not maintained so has bit
rotted. It no longer builds and is missing fixes that have been added to
the C TAP tests.  Remove the old program in preference to the maintained
C TAP tests in tests/util/ktime-t.c.

Also remove the src/util/test/dtest.c program which was an simple
interactive ktime_DateToLong() test.  This program did not even have a
target in the Makefile.

Change-Id: Ib7eaff3c9eec748356101076c0bbc03552e8fb1b
Reviewed-on: https://gerrit.openafs.org/16307
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>