DARWIN: Correct size of arm64 LWP jmp_buf_type to 64-bits
Since the introduction of arm64 support, the LWP jmp_buf_type has been
32 bits.  This causes LWP programs to fail whenever an LWP stack happens
to be allocated above the 4G line:

  stackcheck = 50462976: stack = 50462976
  topstack = 0x50013ff0: stackptr = 0x150008000: stacksize = 0xc000
  Fri Mar 14 10:22:04 2025 LWP: stack overflow in process IO MANAGER!
  from Overflow_Complain()
  Abort trap: 6

This is due to 32-bit truncation of the LWP context topstack value, as
seen in the output above: topstack is reported as 0x50013ff0, but it
should be 0x150013ff0.

Define the size of jmp_buf_type to be uint64 for AFS_ARM64_DARWIN_ENV.

Note: This problem is most frequently observed in cmdebug and bos.

Change-Id: I0203d8f247185ae9414db7753bc8883d7e136721
Reviewed-on: https://gerrit.openafs.org/16343
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>