doc: Rename merge-pod to generate-pod
Rename the merge-pod tool to generate-pod to make it clear the .pod
files are generated, to be consistently named with the generate-man and
generate-html tools, and to allow for additional templating features in
the future.

Change the references of merge-pod to generate-pod in regen.sh,
doc/man-pages/NTMakefile, and doc/man-pages/generate-man.

Change-Id: If5a9084b40630455000514408c17d8a7a90c518e
Reviewed-on: https://gerrit.openafs.org/16326
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>