doc: Regenerate POD for html and AdminRef
Be sure to always regenerate the POD files before rendering the html
formatted man pages and the AdminRef entries.  The regenerated POD files
are written to the objdir directory.

This fixes the html target in docs/man-pages/Makefile when doing an
objdir style build and ensures the PODs are up to date for the places we
currently need them in the build system (html pages and AdminRef
entries).  (The *roff formatted man pages are still expected to be
present in the source tree before make is run. That can be fixed in a
later commit.)

- Add a generate-pod invocation before running generate-html in
  docs/man-pages/Makefile.in

- Add the --stylesheet option to generate-html so it will find the
  style.css file when doing and objdir build.

- Add generate-pod invocation before running generate-xml in the
  docs/xml/AdminRef/Makefile.in

- Update the path to the POD files in docs/xml/AdminRef/Makefile.in to
  use the objdir path.  (Before this change, we relied on the PODs
  being in the srcdir, either for regen.sh or the doc.tar)

- Check for perl in configure and use the detected perl to run the
  various man page generate-* tools.  Later we can make the targets
  conditional based on the configure check.

Change-Id: I24a0a8907f937b65084983e1087a481ec47f3f5a
Reviewed-on: https://gerrit.openafs.org/16331
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>