doc: Generate man pages with the build system
Historically, man pages were maintained separately from the rest of the
source tree.  This change properly integrates the man page generation
with the build system and removes the man page generation during
`regen.sh`.

The man pages will be generated and installed by the top-level makefile
when the required Perl and Perl modules are detected by `configure`.
The HTML formatted man pages are also generated with this change to
avoid bit rot in the HTML generation code.

Running `configure` to generate the man pages is a new requirement, but
requiring autoconf is not an unreasonable burden, and the common case is
to make man page changes and at the same time as code changes, for
example, when adding new command line options.

Packaging systems currently expect the doc.tar archive file with the
source distribution, so continue to provide that file with the
`build-release` tool, but use the build system to generate the man pages
instead of relying on `regen.sh`.  Packaging scripts can safely continue
to unpack the doc.tar file before building the packages.

- Update `src/cf/doc.m4` to check for Perl and required Perl modules to
  enable man page generation.

- Change the `man-pages` target to be conditional in the top-level
  Makefile, enabling man page generation during the build process when
  Perl and the required modules are present.

- Change `make-release` to build man pages and include them in the
  doc.tar archive as a replacement for the man page generation in
  `regen.sh`.  A new `--no-doc-tarball` option is introduced to create a
  set of source distribution files without a doc.tar archive.

- Add the `man` target to doc/man-pages/Makefile.in to generate the POD
  files and the *roff man pages in the objdir tree.

- Add the man and html targets as prerequisites to the all, dest, and
  install targets in doc/man-pages/Makefile.in

- Remove man page generation from `regen.sh`, as it's now handled by the
  build system.

- Update `src/cf/summary.m4` to include man page generation status in
  the configuration summary.

- Move the man page generate file clean up from the top-level Makefile
  to the doc/man-pages/Makefile.in

Change-Id: I442de4c791dd114095035f2f403d564592e759a8
Reviewed-on: https://gerrit.openafs.org/16346
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>