doc: Avoid file slurping in merge-pod
The merge-pod processes the =include directives by using slurping mode
to print the include file to the generated pod output file.

Unfortunately, that drops the line following the =include. In most cases
this in harmless, since the line following the =include is blank, but in
some cases content follows the =include, which is dropped.

For example, the end of pod8/dafileserver.pod.in contains:

    =head1 SEE ALSO

    =include fragments/fileserver-also.pod
    L<fileserver(8)>,

    =head1 COPYRIGHT

The "L<filesever(8)>," line is dropped, and so is missing in the final
generated man page.

Fix this by avoiding the slurping mode and just read each line of the
include file.

While here, rename the merge_pod() subroutine to render() to avoid the
name confusion with the script itself.

Change-Id: I28ebabe5c40301b77c10b3268f84f33fa4bc5f7a
Reviewed-on: https://gerrit.openafs.org/16328
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>