doc: Add generate-pod variable substitution support
Add support to fill in placeholders with named values to the
generate-pod script.  Placeholders are in the form:

    {{ variable_name }}

where `variable_name` is one or more ASCII digits, lowercase letters,
uppercase letters, or the '_' letter.

The double curly brace syntax was chosen because it is common template
syntax and does not conflict with POD syntax or the content in the POD
files.

- Add the --define option to generate-pod to specify one or more
  variable name values.

- Add the variable substitution for each line of template text processed
  by generate-pod, including lines processed in include files.

- Add a routine to look up the value of a named variable and exit
  generate-pod with an error message indicating the file and line with
  the undefined variable name.

- Remove out-of-date comments at the top of generate-pod

Change-Id: I60a15feca52430c787dacb1070c48846dfb114ae
Reviewed-on: https://gerrit.openafs.org/16349
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>