ptserver: Don't use rxgk by default in ptclient
Commit 316b862af6 (ptclient: rxgk support) changed ptclient to use rxgk
if the -rxgk argument was given. But if -rxgk is _not_ given, we default
to passing an rxgk level of 0 to pr_Initialize2, which is a valid rxgk
level (RXGK_LEVEL_CLEAR). This results in ptclient failing with an error
for the default behavior:

    $ ptclient
    Using CellServDB file in /opt/openafs/etc/openafs/
    Making unauthenticated connection to prserver
    libprot: Cannot use rxgk with non-localauth right now
    ptclient: Invalid argument Couldn't initialize protection library

Change the default ptclient rxgk level to RXGK_LEVEL_BOGUS, which is
how other tools like 'pts' work, so we don't try to use rxgk unless the
-rxgk option is given.

Change-Id: I7eb600fecf85f5f90d673742d8b10edce9f080ff
Reviewed-on: https://gerrit.openafs.org/16356
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>