doc: Fix man pages install target for kauth
Correct two 'install' target errors in the doc/man-pages/Makefile.in
when the tree in configured with --enable-kauth, enabling the
installation of the obsolete kauth man pages.

The commit: 'doc: Generate PODs with installation paths' (9e7e727527355)
retired the install-man script but missed updating the commands
to install the kauth man pages.

The commit: 'doc: Do not install *.krb symlinks when kauth is disabled'
(c880fd893560) missed adding a path separator ('/') in the klog.1
symlink target.

No changes are needed for the 'dest' target. That target was not missed
by those commits.

Change-Id: Ie8ad318d4a921ba6947de6319839e445baed3ab1
Reviewed-on: https://gerrit.openafs.org/16360
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>