ktime: Add day validation to ktime_DisplayString()
The ktime_DisplayString() function will crash when the given 'day' has a
value outside the valid range of 0 to 6.  Add an check to ensure the
given 'day' is in range, and if not, return EINVAL instead of crashing.

Add unit tests to check out of range day input values.

The ktime_DisplayString() function is generally unsafe and should be
avoided in new code, since it does not check the input buffer length.
Use of this function in within OpenAFS is currently limited to the 'bos
getrestart' command in the bos client.

Change-Id: Icf38165024c20cc9a418b92f4677b1fd7b45d0eb
Reviewed-on: https://gerrit.openafs.org/16362
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>