ktime: Add day validation to ktime_DisplayString()
The ktime_DisplayString() function will crash when the given 'day' has a
value outside the valid range of 0 to 6.  Add an check to ensure the
given 'day' is in range, and if not, return EINVAL instead of crashing.

Add unit tests to check out of range day input values.

The ktime_DisplayString() function is generally unsafe and should be
avoided in new code, since it does not check the input buffer length.
Use of this function in within OpenAFS is currently limited to the 'bos
getrestart' command in the bos client.

Reviewed-on: https://gerrit.openafs.org/16362
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
(cherry picked from commit 39f7a56e334ff8c2ae2116aa71b9f17dde18cc94)

[1.8: This does not contain the changes to tests/util/ktime-t.c from
master, since 1.8 does not contain relevant prereq changes to the unit
tests.]

Change-Id: I41c46405701babd4e0a6189d864f13b642541aeb
Reviewed-on: https://gerrit.openafs.org/16442
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>