doc: Restore leading '+' in man-pages rule
Commit 168badb037 (build: Remove doc directory checks) accidentally
removed the leading '+' in the commands for the 'man-pages' top-level
make target. The leading '+' causes various 'make' options to get passed
to the sub-make and affects running with -j, -n, -q, or -t, and leaving
it out can cause the following warning for parallel builds:

    cd doc/man-pages && make all
    make[3]: Entering directory `.../doc/man-pages'
    make[3]: warning: jobserver unavailable: using -j1.  Add `+' to parent make rule.

Restore the missing '+' to get rid of the warning and let parallel
builds work properly in that directory. Currently, that directory
doesn't allow for much parallelism, but that could change in the future.

Change-Id: I5a728cdbe4c326f45e2d5a467149bacf619d28d1
Reviewed-on: https://gerrit.openafs.org/16366
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>