lwp: Replace the typedefed bool datatype with int
A new C standard (C23) is coming out and newer versions of C compilers
(e.g. gcc 15) are using the new standard (or an extension of that
standard) as their default.  The C23 standard has added a native C data
type `bool`.  This causes a build error when code tries to use
a typedef to define a `bool` data type:

 iomgr.c:50:23: error: ‘bool’ cannot be defined via ‘typedef’
   50 | typedef unsigned char bool;
      |                       ^~~~
 iomgr.c:50:23: note: ‘bool’ is a keyword with ‘-std=c23’ onwards
 iomgr.c:50:1: warning: useless type name in empty declaration
   50 | typedef unsigned char bool;

There are just a few locations where `bool` is used as a data type.
Replace the use of the `bool` data type with an `int`.

Note: This problem occurred on Fedora 42 which comes with gcc 15.

Reviewed-on: https://gerrit.openafs.org/16369
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
(cherry picked from commit bf64c84058493ee6fe2b1b2fb27b73787aaed318)

Change-Id: Ib959f8c72189df0a797b905388bdd33627ca1ca1
Reviewed-on: https://gerrit.openafs.org/16379
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>