vol: Check response length in fssync-debug
Many places in fssync-debug currently blindly copy struct data from
res.payload.buf (the buffer containing the server's response), without
checking how much data was actually in there. If the server sent more
data than expected we just ignore any extra; but if the server sent too
little data, part of the relevant struct may contain uninitialized
garbage data.

This currently cannot happen, but is fragile, and may happen when
developing changes to the FSSYNC operations. To avoid confusion in such
situations, add a check to verify that the server responded with the
expected number of bytes. If the number of bytes doesn't match, log a
warning, and initialize the struct to zeroes so it doesn't contain
garbage.

Change-Id: Ia83f13794fab635179305f0ff3b77c1f1777c024
Reviewed-on: https://gerrit.openafs.org/16476
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Gaurav Saxena <gsaxena@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>