Remove references to vax/pdp11
VAX and related systems are not exactly supported platforms. Get rid of
any '#ifdef vax' code paths (assume 'vax' is never defined), and do the
same for 'pdp11', 'gould', and 'tahoe' (variants of vax systems).

In src/kauth/crypt.c, this means that MUST_ALIGN is now always defined,
so get rid of MUST_ALIGN and assume it is always set. Also, SALT is now
always 'salt', so get rid of the SALT macro.

In src/afs, this means that AFS_NOSTATS is now never defined. Get rid of
AFS_NOSTATS and assume it is never set.

Change-Id: I4c6d8e3f68e89ac85a40da6ac237e852aeb75d0e
Reviewed-on: https://gerrit.openafs.org/16466
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>