LINUX: Remove AFS_SMP
param.sparc_linux26.h contains some #define workarounds if AFS_SMP and
CONFIG_MODVERSIONS are set. This is unlikely to be needed anymore; this
logic was copied from the original param.sparc_linux22.h param file from
commit 928c5e686b (sparc-linux-20001223).

The AFS_SMP preprocessor symbol is defined by us in our build logic in
MakefileProto.LINUX.in, but whether or not it's defined depends on some
obsolete RedHat-related logic for what kernel variant we build (MP, SP,
EP, etc). Currently, whether we set AFS_SMP is a little inconsistent:

- If --with-linux-kernel-packaging was given, we build for the "SP"
  kernel, and AFS_SMP is not defined.

- If --enable-redhat-buildsys is given, we build for the "Default"
  variant, and AFS_SMP is not defined.

- Otherwise, if CONFIG_SMP is defined we build for the "MP" variant
  (AFS_SMP is defined), or else we build for the "SP" variant (AFS_SMP
  is not defined).

Since this is workaround is unlikely to be needed anymore and doesn't
seem to be applied consistently, get rid of this usage of AFS_SMP. This
is the last user of AFS_SMP (various other users of AFS_SMP have been
removed by various commits), so also change MakefileProto.LINUX.in to
stop defining AFS_SMP.

Change-Id: Icec4cb9531026ca09a736350d616419f653912c8
Reviewed-on: https://gerrit.openafs.org/16384
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>