afs: Remove unused 'now' var in afs_mkdir
afs_mkdir() declares a 'now' timestamp, which it adds to
CallBack.ExpirationTime. But the CallBack data is not currently used in
afs_mkdir(), so this 'now' timestamp is effectively unused. Remove the
useless var.

Move the declaration of 'CallBack' to ensure that the struct isn't used
outside of the RXAFS_MakeDir() call.

Change-Id: I9431ff3a6bbd07bc4dd47078a421cecceb1bc728
Reviewed-on: https://gerrit.openafs.org/16467
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>