afs: Pass relative cb time to CBHash in afs_create
Currently, after calling an RPC that returns a CallBack promise, various
vnode operations call something like this to queue the callback
information locally:

    afs_QueueCallback(tvc, CBHash(CallBack.ExpirationTime), volp);

CBHash takes a relative time in seconds, and converts it to 128-second
units, for use with our local 128-second buckets for tracking callback
timeouts.

Most callers just pass the CallBack.ExpirationTime returned from the
relevant RPC. But in afs_create() since OpenAFS 1.0, we add the 'now'
timestamp (the time when we started calling the RPC) to
CallBack.ExpirationTime, making it an absolute timestamp, causing us to
queue the callback in an arbitrary bucket. If we queue it in a bucket
further in the future than it's supposed to be, we won't clear the
CStatd flag on the vcache until much later than we're supposed to.

For example:

Say we create a file at time 1748960984, and we get an ExpirationTime of
14516 (about 4 hours). In afs_create(), tvc->cbExpires is correctly
calculated to be 1748975500, but CBHash is given 1748975500, not 14516.
So afs_QueueCallBack is given a hash index of 13663871, and so we queue
the callback to bucket 127 (13663871 % CBHTSIZE == 127; assume 'base' is
0). If we gave the correct relative time 14516 to CBHash, we'd get a
bucket of 113.

We won't process the callback expiration (via afs_CheckCallbacks()) until
we get to timeout bucket 127. We advance to the next timeout bucket
about every 127 seconds (via afs_BumpBase()), so that means we won't
process the callback for 16129 seconds. But the callback promise was
only good for 14516 seconds; plus an extra 180 seconds or so for the
viced ServerBias, so about 14696 seconds. That means there is a window
of over 1000 seconds where the fileserver thinks the callback promise
has expired, but the client thinks the callback is still valid. If
someone else modifies the file during that window, the client won't see
a callback break, and so will serve stale data to users.

This situation is perhaps unlikely, since it requires the newly-created
file to be untouched for between about 4 and 4.5 hours (the fileserver
uses different callback times for different file usage, but a
newly-created file can't have any other users, so should always get a
callback of 4 hours). Also, the queued callback gets put in effectively
a random timeout bucket, and we must put it in a timeout bucket further
away than the correct bucket (if we put it in a closer bucket, we'll
just check the expiration time too early, and then move it to the
correct bucket). The correct bucket will be about 113 away, and there
are only 128 buckets. So we have a 113/128 == 88% chance of not
triggering the problem; only a 12% chance that the issue is possible at
the current time.

However, if the fileserver ever returns a shorter callback expiration
time for a newly-created file, this issue could be much more likely to
occur.

To fix this, pass a relative time to the CBHash call for
afs_QueueCallback (the untouched CallBack.ExpirationTime) in
afs_create(), like we do for all other CBHash callers.

Also initialize 'now' to the current time before our do/while retry
loop, to make sure 'now' is initialized. Otherwise, some compilers
complain that 'now' may be used uninitialized, even though this
shouldn't actually be possible:

    .../src/afs/VNOPS/afs_vnop_create.c: In function ‘afs_create’:
    .../src/afs/VNOPS/afs_vnop_create.c:452:48: error: ‘now’ may be used uninitialized in this function [-Werror=maybe-uninitialized]
           tvc->cbExpires = CallBack.ExpirationTime + now;
                            ~~~~~~~~~~~~~~~~~~~~~~~~^~~~~
    cc1: all warnings being treated as errors

Change-Id: Id49e33d1acf50d9a3807911751c4388f3e847322
Reviewed-on: https://gerrit.openafs.org/16601
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>