Make object files depend on generated headers
A few object files currently do not declare a direct Makefile
dependency on some headers they use, and those headers are generated
or installed in a different build step. During a parallel build (or
when using 'make --shuffle'), the header may not be
generated/installed by the time we try to build the object, breaking
the build.

For example:

    $ make --shuffle=reverse
    [...]
      CC  .../src/opr/threadname.lo
    .../src/opr/threadname.c:29:10: fatal error: afs/opr.h: No such file or directory
       29 | #include <afs/opr.h>
          |          ^~~~~~~~~~~
    compilation terminated.

    FAILED COMMAND: /bin/bash ../../libtool --quiet --mode=compile --tag=CC gcc [...] -c .../src/opr/threadname.c

To fix this, make these object files depend on the generated/installed
headers, to ensure that the headers are installed before we build any
objects in opr.

Do this for:

- src/opr, for all object files
- src/rx/simple.example, for sample.cs.o and sample.ss.o
- tests/auth, for keys-t.o

Change-Id: I0d5bfb11d879c33593f545ad1b1bb7b56bc28df4
Reviewed-on: https://gerrit.openafs.org/16391
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>