ktime: Introduce ktime_SetTestTime() for testing
Testing the ktime functions can be challenging since some of them depend
on the system clock.  Introduce a new function to inject a fixed time
value to be used while testing.

Update the ktime_DateToInt32() unit tests to use this new test method
instead of the system clock with a fudge factor.

In preparation to adding more ktime tests, move the ktime_DateToInt32()
tests out of the ktime-t.c main() function to a separate test function.

Fix the integer types in the unit tests to match the types in the
ktime_DateToInt32() function signature.

Introduce the afstest_Scan() macro to iterate over arrays of test case
structures with a for loop.

Change-Id: I86fcdc2dfecd247bfbdb86e06bed5c28be5bf47a
Reviewed-on: https://gerrit.openafs.org/16393
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>