doc: Update command for fetching the gerrit change ID hook
Update the recommended method to download the Gerrit commit-msg git
hook from scp to https.

Instead of scp, provide the URL of the commit-msg git hook and replace
the scp example command with a curl command.

Note that recent openssh versions, in particular those on EL9+, default
to using the sftp protocol and the gerrit server currently doesn't
support sftp.

[mmeffie: switch scp command to https download]

Change-Id: I77d8453086640df4acebef8739041d68b948d27c
Reviewed-on: https://gerrit.openafs.org/16394
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>