afs: Calculate xlator cbExpires more consistently
Most places that set tbc->cbExpires do so like this:

    now = osi_Time();
    RXAFS_Foo(...);
    tvc->cbExpires = callback_expires + now;

That is, the callback expiration time from an RPC is added to the
timestamp of when we started calling the RPC. For example, this is done
in afs_DoBulkStat(), afs_symlink(), and afs_UpdateStatus().

But a couple of places (in afs_LookupVCache() and afs_GetRootVCache())
do this instead (and have since before OpenAFS 1.0):

    RXAFS_Foo(...);
    tvc->cbExpires = callback_expires + osi_Time();

That is, they add the callback expiration time to the current time,
after the RPC has finished running. This doesn't take into account how
long the RPC took to run, and so can result in incorrect callback
expiration times.

Practically speaking, these functions aren't actually called (they are
used for code paths involving the DFS translator). But fix them anyway,
to try to make our various code paths that set 'cbExpires' more
consistent.

Change-Id: Ic28d1417417ccf075c57448909633f0724a3d17f
Reviewed-on: https://gerrit.openafs.org/16602
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: BuildBot <buildbot@rampaginggeek.com>