LINUX: Avoid positive afs_linux_revalidate error
Commit 0ec76649b6 (Minimise crref() and add a fastpath for cache hits),
included in OpenAFS 1.6.0, added an early 'return EIO' to
afs_linux_revalidate(), but this function is supposed to return negative
values on error (which get translated into setting errno by the Linux
kernel).

This means that some syscalls like stat() could theoretically return a
nonsensical value like 5 if this condition is hit, instead of returning
0 on success and -1 on error. This situation shouldn't be possible to
actually hit, though, since 'afs_shuttingdown' should only be changed if
we are shutting down the client, which shouldn't be possible if /afs is
still mounted and afs_linux_revalidate() is getting called. But fix it
anyway, just in case.

To fix this, filter this error through afs_convert_code(), like we do
for all other errors, by jumping to a new label 'out_unlocked'. Fix the
indentation of the 'out' label while we're here, so the labels are
consistent.

Change-Id: Ic9206d286d807cbbb181fdb29c1bed8cc9f4f2f8
Reviewed-on: https://gerrit.openafs.org/16472
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>